backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
reportlab==4.0.7
qrcode[pil]==7.4.2
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.0.0
mercadopago==2.3.0
